requirements.txt
# RedCrowWatch Production Requirements
Flask==3.1.2
opencv-python==4.12.0.88
numpy==2.2.6
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
PyYAML==6.0.2
tweepy==4.16.0
python-dotenv==1.1.1
ultralytics==8.3.202
torch==2.8.0
torchvision==0.23.0
click==8.3.0
tqdm==4.67.1
python-dateutil==2.9.0.post0
librosa==0.11.0
scipy==1.16.2
soundfile==0.13.1
Werkzeug==3.1.3
gunicorn==21.2.0